% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_models.R
\name{comp_models}
\alias{comp_models}
\title{Plot/table hindcast performance}
\usage{
comp_models(dat, comp_age, years = 5, metric = c("md", "mad"))
}
\arguments{
\item{dat}{A data frame with a column providing:\describe{
\item{byr}{brood year}
\item{age#}{one or more columns proving the total return of age # fish using the naming convention age#}
\item{xxx_pred}{one or more columns containing the hindcast prediction for the return of fish of "comp_age" using the naming convention xxx_pred. The prefix "xxx" will be used to name the prediction in the output.}}}

\item{comp_age}{The total age of fish being forecast.}

\item{years}{The number of past years used to evaluate forecast accuracy.}

\item{metric}{The performance metrics used to evaluate forecast accuracy; "md" = "mean deviation", "mad" = "mean absolute deviation" or "mape" = "mean absolute percent error".}
}
\value{
A list containing one figure and one table.
}
\description{
Plot/table hindcast estimates versus actual values and produce a table with hindcast performance metrics.
}
\examples{
dat6 <- prep_brood(deshka, 4:6)
ARIMA6_ar1 <- arima(dat6$age6_ln, order=c(1,0,0))
dat6$ARIMA_pred <- exp(pred_arima(ARIMA6_ar1, x = dat6$age6_ln)[1,])
dat6$es_pred <- pred_es(dat6$age6_ln)
comp_models(dat = dat6, comp_age = 6, years = 5)

}
