% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deshka}
\alias{deshka}
\title{Deshka River Chinook salmon brood table}
\format{A data frame with 44 rows and 7 variables:
\describe{
  \item{bry}{brood year}
  \item{S}{Spawners during each brood year}
  \item{age3}{Age 3 fish returning from each brood year}
  \item{age4}{Age 4 fish returning from each brood year}
  \item{age5}{Age 5 fish returning from each brood year}
  \item{age6}{Age 6 fish returning from each brood year}
  \item{age7}{Age 7 fish returning from each brood year}
}}
\usage{
deshka
}
\description{
Updated thru the 2017 season
}
\keyword{datasets}
