% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot histogram and fitted lognorm/normal distribution}
\usage{
plot_dist(dat)
}
\arguments{
\item{dat}{a vector}
}
\value{
a figure
}
\description{
Use to compart empirical data to fitted distibution.
}
\examples{
plot_dist(brood$age4)

}
