% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_arima.R
\name{pred_arima}
\alias{pred_arima}
\title{Predictions from a ARIMA model}
\usage{
pred_arima(mod, x, xreg = NULL)
}
\arguments{
\item{mod}{a class "lm" object}
}
\value{
a matrix where the first row is the prediction and the second row is the SE of the prediction.
}
\description{
One step forward predictions from a ARIMA model.
}
\examples{
pred_lm(mod)

}
