% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_lm.R
\name{pred_lm}
\alias{pred_lm}
\title{One-step ahead predictions from a linear model}
\usage{
pred_lm(mod)
}
\arguments{
\item{mod}{a class "lm" object}
}
\value{
a matrix where the first row is the prediction and the second row is the SE of the prediction.
}
\description{
Creates a series of one-step ahead predictions for all years beyond the 5th year in the time series using the an exponential smoothing model to be used in evaluating model performance in past seasons.
}
\examples{
dat6 <- prep_brood(deshka, 4:6)
sib6 <- lm(age6_ln ~ age5_ln, data = dat6)
pred_lm(sib6)

}
