% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ma.R
\name{pred_ma}
\alias{pred_ma}
\title{Predictions from a moving average model}
\usage{
pred_ma(var, yrs = NULL)
}
\arguments{
\item{mod}{a class "lm" object}
}
\value{
a vector of predictions.
}
\description{
One step forward predictions from a moving average model
}
\examples{
pred_lm(mod)

}
