% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_brood.R
\name{prep_brood}
\alias{prep_brood}
\title{Prepare a brood table for analysis}
\usage{
prep_brood(brood, age_range)
}
\arguments{
\item{brood}{data frame with the full brood table including columns:\describe{
\item{byr}{brood year}
\item{S}{spawning escapement}
\item{age#}{the total return of age # fish following the naming convention age#}}}

\item{age_range}{a vector with the ages of interest.}
}
\value{
a data frame
}
\description{
Function to select ages of interest from the brood table, remove missing values and calculate logs and log(R/S).
}
\examples{
prep_brood(deshka, 4:6)

}
