% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_models.R
\name{comp_models}
\alias{comp_models}
\title{Plot histogram and fitted lognorm/normal distribution}
\usage{
comp_models(dat, comp_age, years = 5)
}
\arguments{
\item{dat}{a vector}
}
\value{
a figure
}
\description{
Use to compare empirical data to fitted distibution.
}
\examples{
plot_dist(brood$age4)

}
