% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot histogram and fitted lognormal and normal distributions}
\usage{
plot_dist(dat)
}
\arguments{
\item{dat}{a vector}
}
\value{
a figure
}
\description{
Use to compare empirical data to fitted distribution.
}
\examples{
plot_dist(deshka$age5)


}
