% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_es.R
\name{pred_es}
\alias{pred_es}
\title{Predictions from a exponential smoothing model}
\usage{
pred_es(x)
}
\arguments{
\item{mod}{a class "lm" object}
}
\value{
a vector of predictions.
}
\description{
One step forward predictions from a exponential smoothing model
}
\examples{
pred_lm(mod)

}
