% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_lm.R
\name{pred_lm}
\alias{pred_lm}
\title{Predictions from a linear model}
\usage{
pred_lm(mod)
}
\arguments{
\item{mod}{a class "lm" object}
}
\value{
a matrix where the first row is the prediction and the second row is the SE of the prediction.
}
\description{
Leave-one out predictions from a linear model
}
\examples{
pred_lm(mod)

}
