% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ma.R
\name{pred_ma}
\alias{pred_ma}
\title{Predictions from a moving average model}
\usage{
pred_ma(var, yrs = 5)
}
\arguments{
\item{var}{a log transformed time series.}

\item{yrs}{years in the moving average. NULL to average the entire series.}
}
\value{
a matrix with columns "mean" and "median" providing predictions on the natural scale.
}
\description{
One step forward predictions from a moving average model
}
\examples{
pred_ma(var)

}
