% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ma.R
\name{pred_ma}
\alias{pred_ma}
\title{Moving average model estimates}
\usage{
pred_ma(var, yrs = 5)
}
\arguments{
\item{var}{a log transformed time series.}

\item{yrs}{years in the moving average. NULL to average the entire series.}
}
\value{
a matrix with columns "mean" and "median" providing predictions on the natural scale.
}
\description{
Creates a series of one-step ahead predictions for all years beyond the 5th year in the time series using a log scale moving average of the time series to be used in evaluating model performance in past seasons.
}
\examples{
dat6 <- prep_brood(deshka, 4:6)
pred_ma(dat6$age6_ln)

}
